+++ frontend/src/components/ui/StationListWindow/StationSummaryCard.tsx
