+++ frontend/src/components/ui/StationListWindow/fallbacks/EmptyStationsNotice.tsx
